// apps/api/src/comms/enums/comms.enum.ts
export enum MessagePriority {
  NORMAL = 'NORMAL',
  URGENT = 'URGENT',
  FINANCIAL = 'FINANCIAL',
  ACADEMIC = 'ACADEMIC',
}

export enum TriageLabel {
  FEE_QUERY = 'FEE_QUERY',
  ACADEMIC_CONCERN = 'ACADEMIC_CONCERN',
  DISCIPLINARY = 'DISCIPLINARY',
  EMERGENCY = 'EMERGENCY',
  GENERAL_INQUIRY = 'GENERAL_INQUIRY',
  COMPLAINT = 'COMPLAINT',
  PRAISE = 'PRAISE',
}